meditracker-api/requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
sqlalchemy>=2.0.0,<3.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.0.0,<3.0
email-validator>=2.0.0,<3.0
openai>=1.0.0,<2.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.0.0,<5.0
python-jose[cryptography]>=3.3.0,<4.0
pypdf>=4.0.0,<6.0
httpx>=0.27.0,<1.0
